requirements.txt
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-crispy-forms==1.7.2
django-import-export==1.2.0
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
sqlparse==0.3.0
stripe==2.28.0
tablib==0.13.0
urllib3==1.24.3
xlrd==1.2.0
xlwt==1.3.0
